GG-288 Remove abstractions and code that only existed due to Kickstart-related functionality.
