+++ graphitron-codegen-parent/graphitron-java-codegen/src/main/java/no/sikt/graphitron/generators/abstractions/AbstractClassGenerator.java
@@ -1,10 +1,10 @@
 package no.sikt.graphitron.generators.abstractions;
 
+import no.sikt.graphitron.generators.codebuilding.TypeNameFormat;
+import no.sikt.graphitron.generators.dependencies.ServiceDependency;
 import no.sikt.graphitron.javapoet.ClassName;
 import no.sikt.graphitron.javapoet.JavaFile;
 import no.sikt.graphitron.javapoet.TypeSpec;
-import no.sikt.graphitron.generators.codebuilding.TypeNameFormat;
-import no.sikt.graphitron.generators.dependencies.ServiceDependency;
 
 import javax.lang.model.element.Modifier;
 import java.io.File;
