+++ graphitron-codegen-parent/graphitron-java-codegen/src/main/java/no/sikt/graphitron/generators/abstractions/DataFetcherClassGenerator.java
@@ -1,10 +1,10 @@
 package no.sikt.graphitron.generators.abstractions;
 
-import no.sikt.graphitron.javapoet.ClassName;
-import no.sikt.graphitron.javapoet.TypeSpec;
 import no.sikt.graphitron.definitions.interfaces.GenerationTarget;
 import no.sikt.graphitron.generators.codeinterface.wiring.ClassWiringContainer;
 import no.sikt.graphitron.generators.codeinterface.wiring.WiringContainer;
+import no.sikt.graphitron.javapoet.ClassName;
+import no.sikt.graphitron.javapoet.TypeSpec;
 import no.sikt.graphql.schema.ProcessedSchema;
 
 import java.util.ArrayList;
@@ -12,7 +12,7 @@ import java.util.List;
 
 import static no.sikt.graphitron.generators.codeinterface.wiring.WiringContainer.asClassWiring;
 
-abstract public class DataFetcherClassGenerator<T extends GenerationTarget> extends ResolverClassGenerator<T> {
+abstract public class DataFetcherClassGenerator<T extends GenerationTarget> extends AbstractSchemaClassGenerator<T> {
     public static final String DEFAULT_SAVE_DIRECTORY_NAME = "resolvers", FILE_NAME_SUFFIX = "GeneratedDataFetcher";
     protected final List<ClassWiringContainer> fetcherWiringContainer = new ArrayList<>();
