+++ graphitron-codegen-parent/graphitron-java-codegen/src/main/java/no/sikt/graphitron/generators/context/MapperContext.java
@@ -187,10 +187,6 @@ public class MapperContext {
         return targetType;
     }
 
-    public ProcessedSchema getSchema() {
-        return schema;
-    }
-
     public boolean hasRecordReference() {
         return hasRecordReference;
     }
