+++ graphitron-codegen-parent/graphitron-java-codegen/src/main/java/no/sikt/graphitron/generators/dto/EnumDTOGenerator.java
@@ -1,25 +1,22 @@
 package no.sikt.graphitron.generators.dto;
 
-import com.apollographql.federation.graphqljava.Federation;
-import no.sikt.graphitron.javapoet.TypeSpec;
 import no.sikt.graphitron.definitions.objects.EnumDefinition;
+import no.sikt.graphitron.javapoet.TypeSpec;
 import no.sikt.graphql.schema.ProcessedSchema;
 
 import javax.lang.model.element.Modifier;
 import java.util.List;
 
-public class EnumDTOGenerator extends DTOGenerator {
+public class EnumDTOGenerator extends DTOGenerator<EnumDefinition> {
 
     public EnumDTOGenerator(ProcessedSchema processedSchema) {
         super(processedSchema);
     }
 
-    protected TypeSpec generate(EnumDefinition enumDefinition) {
+    @Override
+    public TypeSpec generate(EnumDefinition enumDefinition) {
         var enumBuilder = TypeSpec.enumBuilder(enumDefinition.getName()).addModifiers(Modifier.PUBLIC);
-        enumDefinition.getFields().forEach(field -> {
-            enumBuilder.addEnumConstant(field.getName());
-        });
-
+        enumDefinition.getFields().forEach(field -> enumBuilder.addEnumConstant(field.getName()));
         return enumBuilder.build();
     }
