+++ graphitron-codegen-parent/graphitron-java-codegen/src/main/java/no/sikt/graphitron/generators/dto/InputDTOGenerator.java
@@ -1,18 +1,18 @@
 package no.sikt.graphitron.generators.dto;
 
-import no.sikt.graphitron.javapoet.TypeSpec;
 import no.sikt.graphitron.definitions.objects.InputDefinition;
+import no.sikt.graphitron.javapoet.TypeSpec;
 import no.sikt.graphql.schema.ProcessedSchema;
 
 import java.util.List;
 
-public class InputDTOGenerator extends DTOGenerator {
-
+public class InputDTOGenerator extends DTOGenerator<InputDefinition> {
     public InputDTOGenerator(ProcessedSchema processedSchema) {
         super(processedSchema);
     }
 
-    protected TypeSpec generate(InputDefinition target) {
+    @Override
+    public TypeSpec generate(InputDefinition target) {
         return getTypeSpecBuilder(target.getName(), target.getFields()).build();
     }
