+++ graphitron-codegen-parent/graphitron-java-codegen/src/main/java/no/sikt/graphitron/generators/dto/UnionDTOGenerator.java
@@ -1,19 +1,20 @@
 package no.sikt.graphitron.generators.dto;
 
-import no.sikt.graphitron.javapoet.TypeSpec;
 import no.sikt.graphitron.definitions.objects.UnionDefinition;
+import no.sikt.graphitron.javapoet.TypeSpec;
 import no.sikt.graphql.schema.ProcessedSchema;
 
 import javax.lang.model.element.Modifier;
 import java.util.List;
 
-public class UnionDTOGenerator extends DTOGenerator {
+public class UnionDTOGenerator extends DTOGenerator<UnionDefinition> {
 
     public UnionDTOGenerator(ProcessedSchema processedSchema) {
         super(processedSchema);
     }
 
-    protected TypeSpec generate(UnionDefinition target) {
+    @Override
+    public TypeSpec generate(UnionDefinition target) {
         return TypeSpec.interfaceBuilder(target.getName()).addModifiers(Modifier.PUBLIC).build();
     }
