+++ graphitron-codegen-parent/graphitron-java-codegen/src/test/resources/datafetchers/fetch/pagination/operation/default/expected/QueryGeneratedDataFetcher.java
@@ -1,9 +1,7 @@
 package fake.code.generated.resolvers.query;
 
 import fake.code.generated.queries.QueryDBQueries;
-import fake.graphql.example.model.DummyConnection;
-import fake.graphql.example.model.DummyConnectionEdge;
-import fake.graphql.example.model.PageInfo;
+import fake.graphql.example.model.DummyType;
 import graphql.schema.DataFetcher;
 import java.lang.Integer;
 import java.lang.String;
@@ -11,9 +9,10 @@ import java.util.concurrent.CompletableFuture;
 
 import no.sikt.graphql.helpers.resolvers.DataFetcherHelper;
 import no.sikt.graphql.helpers.resolvers.ResolverHelpers;
+import no.sikt.graphql.relay.ConnectionImpl;
 
 public class QueryGeneratedDataFetcher {
-    public static DataFetcher<CompletableFuture<DummyConnection>> query() {
+    public static DataFetcher<CompletableFuture<ConnectionImpl<DummyType>>> query() {
         return env -> {
             Integer first = env.getArgument("first");
             String after = env.getArgument("after");
@@ -21,13 +20,7 @@ public class QueryGeneratedDataFetcher {
             return new DataFetcherHelper(env).loadPaginated(
                     pageSize, 1000,
                     (ctx, selectionSet) -> QueryDBQueries.queryForQuery(ctx, pageSize,after, selectionSet),
-                    (ctx, resolverKeys) -> QueryDBQueries.countQueryForQuery(ctx),
-                    (connection) ->  {
-                        var edges = connection.getEdges().stream().map(it -> new DummyConnectionEdge(it.getCursor() == null ? null : it.getCursor().getValue(), it.getNode())).toList();
-                        var page = connection.getPageInfo();
-                        var graphPage = new PageInfo(page.isHasPreviousPage(), page.isHasNextPage(), page.getStartCursor() == null ? null : page.getStartCursor().getValue(), page.getEndCursor() == null ? null : page.getEndCursor().getValue());
-                        return new DummyConnection(edges, graphPage, connection.getNodes(), connection.getTotalCount());
-                    }
+                    (ctx, resolverKeys) -> QueryDBQueries.countQueryForQuery(ctx)
             );
         };
     }
