+++ graphitron-codegen-parent/graphitron-java-codegen/src/test/resources/datafetchers/fetch/pagination/splitquery/default/expected/WrapperGeneratedDataFetcher.java
@@ -1,9 +1,7 @@
 package fake.code.generated.resolvers.query;
 
 import fake.code.generated.queries.WrapperDBQueries;
-import fake.graphql.example.model.DummyConnection;
-import fake.graphql.example.model.DummyConnectionEdge;
-import fake.graphql.example.model.PageInfo;
+import fake.graphql.example.model.DummyType;
 import fake.graphql.example.model.Wrapper;
 import graphql.schema.DataFetcher;
 import java.lang.Integer;
@@ -12,9 +10,10 @@ import java.util.concurrent.CompletableFuture;
 
 import no.sikt.graphql.helpers.resolvers.DataFetcherHelper;
 import no.sikt.graphql.helpers.resolvers.ResolverHelpers;
+import no.sikt.graphql.relay.ConnectionImpl;
 
 public class WrapperGeneratedDataFetcher {
-    public static DataFetcher<CompletableFuture<DummyConnection>> query() {
+    public static DataFetcher<CompletableFuture<ConnectionImpl<DummyType>>> query() {
         return env -> {
             Wrapper wrapper = env.getSource();
             Integer first = env.getArgument("first");
@@ -23,13 +22,7 @@ public class WrapperGeneratedDataFetcher {
             return new DataFetcherHelper(env).loadPaginated(
                     wrapper.getQueryKey(), pageSize, 1000,
                     (ctx, resolverKeys, selectionSet) -> WrapperDBQueries.queryForWrapper(ctx, resolverKeys, pageSize, after, selectionSet),
-                    (ctx, resolverKeys) -> WrapperDBQueries.countQueryForWrapper(ctx, resolverKeys),
-                    (connection) ->  {
-                        var edges = connection.getEdges().stream().map(it -> new DummyConnectionEdge(it.getCursor() == null ? null : it.getCursor().getValue(), it.getNode())).toList();
-                        var page = connection.getPageInfo();
-                        var graphPage = new PageInfo(page.isHasPreviousPage(), page.isHasNextPage(), page.getStartCursor() == null ? null : page.getStartCursor().getValue(), page.getEndCursor() == null ? null : page.getEndCursor().getValue());
-                        return new DummyConnection(edges, graphPage, connection.getNodes(), connection.getTotalCount());
-                    }
+                    (ctx, resolverKeys) -> WrapperDBQueries.countQueryForWrapper(ctx, resolverKeys)
             );
         };
     }
