+++ graphitron-codegen-parent/graphitron-java-codegen/src/test/resources/datafetchers/fetch/services/operation/withPagination/expected/QueryGeneratedDataFetcher.java
@@ -1,9 +1,7 @@
 package fake.code.generated.resolvers.query;
 
 import fake.code.generated.transform.RecordTransformer;
-import fake.graphql.example.model.CustomerConnection;
-import fake.graphql.example.model.CustomerConnectionEdge;
-import fake.graphql.example.model.PageInfo;
+import fake.graphql.example.model.CustomerTable;
 import graphql.schema.DataFetcher;
 import java.lang.Integer;
 import java.lang.String;
@@ -12,9 +10,10 @@ import java.util.concurrent.CompletableFuture;
 import no.sikt.graphitron.codereferences.services.ResolverFetchService;
 import no.sikt.graphql.helpers.resolvers.ResolverHelpers;
 import no.sikt.graphql.helpers.resolvers.ServiceDataFetcherHelper;
+import no.sikt.graphql.relay.ConnectionImpl;
 
 public class QueryGeneratedDataFetcher {
-    public static DataFetcher<CompletableFuture<CustomerConnection>> query() {
+    public static DataFetcher<CompletableFuture<ConnectionImpl<CustomerTable>>> query() {
         return env -> {
             Integer first = env.getArgument("first");
             String after = env.getArgument("after");
@@ -25,13 +24,7 @@ public class QueryGeneratedDataFetcher {
                     pageSize, 1000,
                     () -> resolverFetchService.queryList(pageSize, after),
                     (resolverKeys) -> resolverFetchService.countQueryList(),
-                    (recordTransform, response) -> recordTransform.customerTableRecordToGraphType(response, ""),
-                    (connection) ->  {
-                        var edges = connection.getEdges().stream().map(it -> new CustomerConnectionEdge(it.getCursor() == null ? null : it.getCursor().getValue(), it.getNode())).toList();
-                        var page = connection.getPageInfo();
-                        var graphPage = new PageInfo(page.isHasPreviousPage(), page.isHasNextPage(), page.getStartCursor() == null ? null : page.getStartCursor().getValue(), page.getEndCursor() == null ? null : page.getEndCursor().getValue());
-                        return new CustomerConnection(edges, graphPage, connection.getNodes(), connection.getTotalCount());
-                    }
+                    (recordTransform, response) -> recordTransform.customerTableRecordToGraphType(response, "")
             );
         };
     }
