+++ graphitron-codegen-parent/graphitron-java-codegen/src/test/resources/datafetchers/fetch/services/splitquery/withPagination/expected/WrapperGeneratedDataFetcher.java
@@ -1,9 +1,7 @@
 package fake.code.generated.resolvers.query;
 
 import fake.code.generated.transform.RecordTransformer;
-import fake.graphql.example.model.CustomerConnection;
-import fake.graphql.example.model.CustomerConnectionEdge;
-import fake.graphql.example.model.PageInfo;
+import fake.graphql.example.model.CustomerTable;
 import fake.graphql.example.model.Wrapper;
 import graphql.schema.DataFetcher;
 import java.lang.Integer;
@@ -13,9 +11,10 @@ import java.util.concurrent.CompletableFuture;
 import no.sikt.graphitron.codereferences.services.ResolverFetchService;
 import no.sikt.graphql.helpers.resolvers.ResolverHelpers;
 import no.sikt.graphql.helpers.resolvers.ServiceDataFetcherHelper;
+import no.sikt.graphql.relay.ConnectionImpl;
 
 public class WrapperGeneratedDataFetcher {
-    public static DataFetcher<CompletableFuture<CustomerConnection>> query() {
+    public static DataFetcher<CompletableFuture<ConnectionImpl<CustomerTable>>> query() {
         return env -> {
             Wrapper wrapper = env.getSource();
             Integer first = env.getArgument("first");
@@ -27,13 +26,7 @@ public class WrapperGeneratedDataFetcher {
                     wrapper.getQueryKey(), pageSize, 1000,
                     (resolverKeys) -> resolverFetchService.queryMap(resolverKeys, pageSize, after),
                     (resolverKeys) -> resolverFetchService.countQueryMap(resolverKeys),
-                    (recordTransform, response) -> recordTransform.customerTableRecordToGraphType(response, ""),
-                    (connection) ->  {
-                        var edges = connection.getEdges().stream().map(it -> new CustomerConnectionEdge(it.getCursor() == null ? null : it.getCursor().getValue(), it.getNode())).toList();
-                        var page = connection.getPageInfo();
-                        var graphPage = new PageInfo(page.isHasPreviousPage(), page.isHasNextPage(), page.getStartCursor() == null ? null : page.getStartCursor().getValue(), page.getEndCursor() == null ? null : page.getEndCursor().getValue());
-                        return new CustomerConnection(edges, graphPage, connection.getNodes(), connection.getTotalCount());
-                    }
+                    (recordTransform, response) -> recordTransform.customerTableRecordToGraphType(response, "")
             );
         };
     }
