--- graphitron-codegen-parent/graphitron-java-codegen/src/test/resources/dto/type/unorderedConnectionType/schema.graphqls
@@ -1,11 +1,0 @@
-type DummyConnection {
-    totalCount: Int
-    pageInfo: PageInfo
-    nodes: [DummyType!]!
-    edges: [DummyConnectionEdge]
-}
-
-type DummyConnectionEdge {
-    cursor: String
-    node: DummyType
-}
