--- graphitron-codegen-parent/graphitron-java-codegen/src/test/resources/dto/type/unorderedPageInfo/schema.graphqls
@@ -1,6 +1,0 @@
-type PageInfo {
-    startCursor: String
-    hasPreviousPage: Boolean!
-    hasNextPage: Boolean!
-    endCursor: String
-}
